requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pandas>=2.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
python-dotenv>=1.0.0
